Refactor CreateGameForm
